PlaybackTracker: Extract Metdata -> Scrobble to Scrobble.kt
